review-runner/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# LLM Integration
openai==1.3.0
anthropic==0.7.0

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.40

# Utilities
tenacity==8.2.3
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
